requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
defusedxml==0.7.1
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
django-health-check==3.16.5
djangorestframework==3.13.1
gunicorn==20.1.0
httpie==3.0.2
idna==3.3
lxml==4.7.1
multidict==6.0.2
Pygments==2.11.2
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
